+++ muxmiddleware_test.go
@@ -26,10 +26,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/gofiber/fiber/v2"
 	"github.com/sirupsen/logrus"
 	"github.com/sirupsen/logrus/hooks/test"
-	"github.com/valyala/fasthttp"
 	"gotest.tools/assert"
 )
 
@@ -443,28 +441,6 @@ func TestMuxLogMiddleware(t *testing.T) {
 
 		hook.Reset()
 	})
-}
-
-func TestFiberLogMiddleware(t *testing.T) {
-	app := fiber.New()
-	c := app.AcquireCtx(&fasthttp.RequestCtx{})
-	defer app.ReleaseCtx(c)
-
-	log, hook := test.NewNullLogger()
-
-	app.Use(RequestFiberMiddlewareLogger(log, nil))
-	called := false
-	app.Get("/test", func(c *fiber.Ctx) error {
-		called = true
-		return nil
-	})
-
-	req := httptest.NewRequest("GET", "/test", nil)
-	_, err := app.Test(req)
-	assert.NilError(t, err)
-
-	assert.Equal(t, called, true)
-	t.Log(hook.AllEntries()[0])
 }
 
 func logAssertions(t *testing.T, logEntry *logrus.Entry, expected ExpectedLogFields) string {
